requirements.txt
moviepy
openai
google-api-python-client
google-auth
google-auth-oauthlib
gtts